Don't filter default package list for first boot dexopt
commit 20a8542b51 ("Implement app downgrading.") added filtering of the
list of packages to dexopt and for REASON_FIRST_BOOT this filtered the
list of packages based on last active time which is determined as the
maximum of a package's last used time and first install time. During
first boot a package's last used time is 0 so the last active time is
based on a package's installation time which itself is based on its mtime.
When filesystem images are built for target_files.zip a fixed file
timestamp (01/01/2009 00:00:00 UTC) is set. This can be a problem during
first boot if pm.dexopt.downgrade_after_inactive_days is set because in this
case packages with a last active time (installation time in this case) more
than pm.dexopt.downgrade_after_inactive_days in the past will be be filtered
out of the default package list and not dexopted so if either, for example,
an update package is flashed or an old enough image from $(PRODUCT_OUT) is
flashed no packages will pass the current filtering so fix this by not doing
any filtering for first boot.

Test: Set pm.dexopt.downgrade_after_inactive_days=10 on
  aosp_cheetah-userdebug and fastboot -w update $OUT/aosp_cheetah-img.zip and
  confirm with debugger and ART Service logging that there are now packages
  being dexopted.
Test: atest ArtServiceTests

Change-Id: I4ebd74766cfe21db4001b44bd292cc042a8e23f0